+++ noxuslab/templates/multi_process/_process_template/__init__.py
@@ -1,1 +1,0 @@
-"""Process `__process__` in workspace `__workspace__`."""